+++ examples/react/.env.development
@@ -10,7 +10,7 @@ IRRELEVANT_ENV=$DEPEND_ENV
 IRRELEVANT_ESCAPE_ENV=irrelevant$
 DEPEND_ENV=depend
 NODE_ENV=development
-FARM_BASE_TEST=12asdasd56
+FARM_BASE_TEST=12aasdasaaasdasddasdsdasd512
