+++ packages/core/src/utils/file.ts
@@ -1,6 +1,3 @@
-import path from 'node:path';
-import { pathToFileURL } from 'node:url';
-import { readFileSync } from 'node:fs';
 interface FileNode {
   isLeaf: boolean;
   name: string;
@@ -70,50 +67,4 @@ export function generateFileTreeHtml(node: FileNode[]): string {
         </body>
       </html>
       `;
-}
-
-export function getDependenciesRecursive(config: any) {
-  const content = readFileSync(config.resolveConfigPath, 'utf-8');
-  const dependencyRegex = /import\s.*?from\s['"](.+?)['"]/g;
-  const requireRegex = /require\s*\(\s*['"](.+?)['"]\s*\)/g;
-  const allDependencies = [];
-  const dependencies = [];
-
-  let match;
-  while ((match = dependencyRegex.exec(content)) !== null) {
-    dependencies.push(match[1]);
-  }
-
-  while ((match = requireRegex.exec(content)) !== null) {
-    dependencies.push(match[1]);
-  }
-
-  for (const dependency of dependencies) {
-    const dependencyPath = path.resolve(
-      path.dirname(config.filePath),
-      dependency
-    );
-    // 检查依赖项是否在项目内部，而不是在node_modules中
-    if (!dependencyPath.includes('node_modules')) {
-      allDependencies.push(dependencyPath);
-      getDependenciesRecursive(dependencyPath);
-    }
-  }
-
-  return allDependencies;
-}
-
-export function isInternalDependency(dependencyPath: string) {
-  const projectRoot = path.resolve(__dirname);
-  return dependencyPath.startsWith(projectRoot);
-}
-
-export async function importFresh(modulePath: string) {
-  const cacheBustingModulePath = `${modulePath}?update=${Date.now()}`;
-  if (process.platform === 'win32') {
-    return (await import(pathToFileURL(cacheBustingModulePath).toString()))
-      .default;
-  } else {
-    return (await import(cacheBustingModulePath)).default;
-  }
 }
