+++ packages/core/src/watcher/index.ts
@@ -7,14 +7,8 @@ import { Compiler } from '../compiler/index.js';
 import { DevServer } from '../server/index.js';
 import { Config, JsFileWatcher } from '../../binding/index.js';
 import { compilerHandler, DefaultLogger, clearScreen } from '../utils/index.js';
-import {
-  DEFAULT_HMR_OPTIONS,
-  JsPlugin,
-  normalizeUserCompilationConfig,
-  resolveUserConfig
-} from '../index.js';
+import { DEFAULT_HMR_OPTIONS, resolveCompiler } from '../index.js';
 import { __FARM_GLOBAL__ } from '../config/_global.js';
-import { setProcessEnv } from '../config/env.js';
 
 import type { UserConfig } from '../config/index.js';
 
@@ -69,22 +63,10 @@ export class FileWatcher implements ImplFileWatcher {
           )} change`
         );
         if (this.serverOrCompiler instanceof DevServer) {
+          console.log(this.serverOrCompiler);
           await this.serverOrCompiler.close();
         }
-        const config: UserConfig = await resolveUserConfig(
-          this.options.inlineConfig,
-          this._logger
-        );
-        const normalizedConfig = await normalizeUserCompilationConfig(config);
-        setProcessEnv(normalizedConfig.config.mode);
-        const compiler = new Compiler(normalizedConfig);
-        const devServer = new DevServer(compiler, this._logger, config);
-        devServer.listen();
-        if (normalizedConfig.config.mode === 'development') {
-          normalizedConfig.jsPlugins.forEach((plugin: JsPlugin) =>
-            plugin.configDevServer?.(devServer)
-          );
-        }
+        resolveCompiler(this.options as any, this._logger);
         return;
       }
       try {
